# docs/swagger.yaml
definitions:
  handlers.APIResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  models.CardMinor:
    properties:
      deck:
        example: 1
        type: integer
      id:
        type: integer
      image:
        description: Full URL
        type: string
      meanings:
        items:
          $ref: '#/definitions/models.MeaningRef'
        type: array
      name:
        example: The Fool
        type: string
      rank:
        example: 10
        type: integer
      suit:
        example: 1
        type: integer
    type: object
  models.CardMinorInput:
    properties:
      deck:
        example: 1
        type: integer
      rank:
        example: 10
        type: integer
      suit:
        example: 1
        type: integer
    type: object
...
  models.IDOnly:
    properties:
      id:
        example: 1
        type: integer
    type: object
...
  models.MeaningMinor:
    properties:
      id:
        type: integer
      meaning:
        type: string
      position:
        $ref: '#/definitions/models.MeaningPosition'
      rank:
        type: integer
      source:
        type: integer
      suit:
        type: integer
    type: object
  models.MeaningMinorInput:
    properties:
      meaning:
        example: Active communication and drive
        type: string
      position:
        allOf:
        - $ref: '#/definitions/models.MeaningPosition'
        example: straight
      rank:
        example: 5
        type: integer
      source:
        example: 1
        type: integer
      suit:
        example: 1
        type: integer
    type: object
  models.MeaningPosition:
    enum:
    - straight
    - reverted
    type: string
    x-enum-varnames:
    - PositionStraight
    - PositionReverted
  models.MeaningRef:
    properties:
      id:
        type: integer
      position:
        type: string
      source:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /cards/minor:
    get:
      consumes:
      - application/json
      description: Returns all minor arcana cards from the specified deck
      parameters:
      - description: Deck ID (required)
        in: query
        name: deckId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CardMinor'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: List Minor Arcana cards
      tags:
      - cards
    post:
      consumes:
      - application/json
      description: Adds a new card of interpretations
      parameters:
      - description: CardMinor data
        in: body
        name: card
        required: true
        schema:
          $ref: '#/definitions/models.CardMinorInput'
      produces:
      - application/json
      responses:
        "201":
          description: Returns the ID of the created card
          schema:
            $ref: '#/definitions/models.IDOnly'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Create a new card
      tags:
      - cards
  /cards/minor/{id}:
    delete:
      description: Deletes a card by ID
      parameters:
      - description: CardMinor ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Delete a card
      tags:
      - cards
    get:
      description: Retrieves a card by its ID
      parameters:
      - description: CardMinor ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CardMinor'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Get card by ID
      tags:
      - cards
    put:
      consumes:
      - application/json
      description: Updates an existing card
      parameters:
      - description: CardMinor ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated card
        in: body
        name: card
        required: true
        schema:
          $ref: '#/definitions/models.CardMinorInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CardMinor'
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Update a card
      tags:
      - cards
...
  /meanings/minor:
    get:
      consumes:
      - application/json
      description: Returns a list of meanings for minor arcana cards with optional
        filters
      parameters:
      - description: Card number (optional)
        in: query
        name: number
        type: integer
      - description: Card position
        enum:
        - straight
        - reverted
        in: query
        name: position
        type: string
      - description: Source ID (optional)
        in: query
        name: source
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MeaningMinor'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Get Minor Arcana meanings
      tags:
      - meanings
    post:
      consumes:
      - application/json
      description: Adds a new MinorMeaning of interpretations
      parameters:
      - description: MinorMeaning data
        in: body
        name: MinorMeaning
        required: true
        schema:
          $ref: '#/definitions/models.MeaningMinorInput'
      produces:
      - application/json
      responses:
        "201":
          description: Returns the ID of the created MinorMeaning
          schema:
            $ref: '#/definitions/models.IDOnly'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Create a new MinorMeaning
      tags:
      - meanings
  /meanings/minor/{id}:
    delete:
      description: Deletes a MinorMeaning by ID
      parameters:
      - description: MinorMeaning ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Delete a MinorMeaning
      tags:
      - meanings
    get:
      description: Retrieves a MinorMeaning by its ID
      parameters:
      - description: MinorMeaning ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MeaningMinor'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Get MinorMeaning by ID
      tags:
      - meanings
    put:
      consumes:
      - application/json
      description: Updates an existing MinorMeaning
      parameters:
      - description: MinorMeaning ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated MinorMeaning
        in: body
        name: MinorMeaning
        required: true
        schema:
          $ref: '#/definitions/models.MeaningMinorInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MeaningMinor'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Update a MinorMeaning
      tags:
      - meanings
...
swagger: "2.0"